Clang cleanup

+++ src/swganh_core/simulation/node_unittest.cc
@@ -47,10 +47,10 @@ BOOST_AUTO_TEST_CASE(CanInsertRemoveObject)
 	obj->UpdateAABB();
 
 	root_node_.InsertObject(obj);
-	BOOST_CHECK_EQUAL(1, root_node_.GetContainedObjects().size());
+	BOOST_CHECK_EQUAL(size_t(1), root_node_.GetContainedObjects().size());
 
 	root_node_.RemoveObject(obj);
-	BOOST_CHECK_EQUAL(0, root_node_.GetContainedObjects().size());
+	BOOST_CHECK_EQUAL(size_t(0), root_node_.GetContainedObjects().size());
 }
 
 ///
@@ -90,15 +90,15 @@ BOOST_AUTO_TEST_CASE(VerifyQuadrantSplit)
 	root_node_.InsertObject(obj3);
 	root_node_.InsertObject(obj4);
 	
-	BOOST_CHECK_EQUAL(1, root_node_.GetLeafNodes()[NW_QUADRANT]->GetObjects().size());
-	BOOST_CHECK_EQUAL(1, root_node_.GetLeafNodes()[NE_QUADRANT]->GetObjects().size());
-	BOOST_CHECK_EQUAL(1, root_node_.GetLeafNodes()[SW_QUADRANT]->GetObjects().size());
-	BOOST_CHECK_EQUAL(1, root_node_.GetLeafNodes()[SE_QUADRANT]->GetObjects().size());
+	BOOST_CHECK_EQUAL(size_t(1), root_node_.GetLeafNodes()[NW_QUADRANT]->GetObjects().size());
+	BOOST_CHECK_EQUAL(size_t(1), root_node_.GetLeafNodes()[NE_QUADRANT]->GetObjects().size());
+	BOOST_CHECK_EQUAL(size_t(1), root_node_.GetLeafNodes()[SW_QUADRANT]->GetObjects().size());
+	BOOST_CHECK_EQUAL(size_t(1), root_node_.GetLeafNodes()[SE_QUADRANT]->GetObjects().size());
     
-	BOOST_CHECK_EQUAL(1, (*root_node_.GetLeafNodes()[NW_QUADRANT]->GetObjects().begin())->GetObjectId());
-	BOOST_CHECK_EQUAL(2, (*root_node_.GetLeafNodes()[NE_QUADRANT]->GetObjects().begin())->GetObjectId());
-	BOOST_CHECK_EQUAL(3, (*root_node_.GetLeafNodes()[SW_QUADRANT]->GetObjects().begin())->GetObjectId());
-	BOOST_CHECK_EQUAL(4, (*root_node_.GetLeafNodes()[SE_QUADRANT]->GetObjects().begin())->GetObjectId());
+	BOOST_CHECK_EQUAL(uint64_t(1), (*root_node_.GetLeafNodes()[NW_QUADRANT]->GetObjects().begin())->GetObjectId());
+	BOOST_CHECK_EQUAL(uint64_t(2), (*root_node_.GetLeafNodes()[NE_QUADRANT]->GetObjects().begin())->GetObjectId());
+	BOOST_CHECK_EQUAL(uint64_t(3), (*root_node_.GetLeafNodes()[SW_QUADRANT]->GetObjects().begin())->GetObjectId());
+	BOOST_CHECK_EQUAL(uint64_t(4), (*root_node_.GetLeafNodes()[SE_QUADRANT]->GetObjects().begin())->GetObjectId());
 
 	root_node_.RemoveObject(obj1);
 	root_node_.RemoveObject(obj2);
@@ -120,11 +120,11 @@ BOOST_AUTO_TEST_CASE(CanQuery)
 
 	root_node_.InsertObject(obj);
 
-	BOOST_CHECK_EQUAL(1, root_node_.Query(QueryBox( Point(0.0f, 0.0f), Point(15.0f, 15.0f) )).size());
+	BOOST_CHECK_EQUAL(size_t(1), root_node_.Query(QueryBox( Point(0.0f, 0.0f), Point(15.0f, 15.0f) )).size());
 
 	root_node_.RemoveObject(obj);
 
-	BOOST_CHECK_EQUAL(0, root_node_.Query(QueryBox( Point(0.0f, 0.0f), Point(15.0f, 15.0f) )).size());
+	BOOST_CHECK_EQUAL(size_t(0), root_node_.Query(QueryBox( Point(0.0f, 0.0f), Point(15.0f, 15.0f) )).size());
 }
 
 ///
@@ -139,16 +139,16 @@ BOOST_AUTO_TEST_CASE(CanUpdateObject)
 
 	root_node_.InsertObject(obj);
 
-	BOOST_CHECK_EQUAL(1, root_node_.GetLeafNodes()[NE_QUADRANT]->GetObjects().size());
-	BOOST_CHECK_EQUAL(0, root_node_.GetLeafNodes()[SE_QUADRANT]->GetObjects().size());
+	BOOST_CHECK_EQUAL(size_t(1), root_node_.GetLeafNodes()[NE_QUADRANT]->GetObjects().size());
+	BOOST_CHECK_EQUAL(size_t(0), root_node_.GetLeafNodes()[SE_QUADRANT]->GetObjects().size());
 
 	glm::vec3 new_position(10.0f, 0.0f, -10.0f);
 	auto old_bounding_volume = obj->GetAABB();
 	obj->SetPosition(new_position);
 	root_node_.UpdateObject(obj, old_bounding_volume, obj->GetAABB());
 
-	BOOST_CHECK_EQUAL(0, root_node_.GetLeafNodes()[NE_QUADRANT]->GetObjects().size());
-	BOOST_CHECK_EQUAL(1, root_node_.GetLeafNodes()[SE_QUADRANT]->GetObjects().size());
+	BOOST_CHECK_EQUAL(size_t(0), root_node_.GetLeafNodes()[NE_QUADRANT]->GetObjects().size());
+	BOOST_CHECK_EQUAL(size_t(1), root_node_.GetLeafNodes()[SE_QUADRANT]->GetObjects().size());
 
 	root_node_.RemoveObject(obj);
 }
@@ -172,16 +172,16 @@ BOOST_AUTO_TEST_CASE(CanInsertRemoveQueryOneThousand)
 		root_node_.InsertObject(objects[i]);
 	}
 
-	BOOST_CHECK_EQUAL(1000, root_node_.GetContainedObjects().size());
-	BOOST_CHECK_EQUAL(1000, root_node_.Query(QueryBox(Point(-3000, -3000), Point(3000, 3000))).size());
+	BOOST_CHECK_EQUAL(size_t(1000), root_node_.GetContainedObjects().size());
+	BOOST_CHECK_EQUAL(size_t(1000), root_node_.Query(QueryBox(Point(-3000, -3000), Point(3000, 3000))).size());
 
 	for(int i = 0; i < 1000; i++)
 	{
 		root_node_.RemoveObject(objects[i]);
 	}
 
-	BOOST_CHECK_EQUAL(0, root_node_.GetContainedObjects().size());
-	BOOST_CHECK_EQUAL(0, root_node_.Query(QueryBox(Point(-3000, -3000), Point(3000, 3000))).size());
+	BOOST_CHECK_EQUAL(size_t(0), root_node_.GetContainedObjects().size());
+	BOOST_CHECK_EQUAL(size_t(0), root_node_.Query(QueryBox(Point(-3000, -3000), Point(3000, 3000))).size());
 }
 
 ///
@@ -201,16 +201,16 @@ BOOST_AUTO_TEST_CASE(CanInsertRemoveQueryTenThousand)
 		root_node_.InsertObject(objects[i]);
 	}
 
-	BOOST_CHECK_EQUAL(10000, root_node_.GetContainedObjects().size());
-	BOOST_CHECK_EQUAL(10000, root_node_.Query(QueryBox(Point(-3000, -3000), Point(3000, 3000))).size());
+	BOOST_CHECK_EQUAL(size_t(10000), root_node_.GetContainedObjects().size());
+	BOOST_CHECK_EQUAL(size_t(10000), root_node_.Query(QueryBox(Point(-3000, -3000), Point(3000, 3000))).size());
 
 	for(int i = 0; i < 10000; i++)
 	{
 		root_node_.RemoveObject(objects[i]);
 	}
 
-	BOOST_CHECK_EQUAL(0, root_node_.GetContainedObjects().size());
-	BOOST_CHECK_EQUAL(0, root_node_.Query(QueryBox(Point(-3000, -3000), Point(3000, 3000))).size());
+	BOOST_CHECK_EQUAL(size_t(0), root_node_.GetContainedObjects().size());
+	BOOST_CHECK_EQUAL(size_t(0), root_node_.Query(QueryBox(Point(-3000, -3000), Point(3000, 3000))).size());
 }
 
 BOOST_AUTO_TEST_SUITE_END()
